Konverse_Python/requirements.txt
flask
google.generativeai
pypdf
flask-cors
python-dotenv
flask-restful
gunicorn
convex
pandas
python-docx
inscriptis
